+++ themes-default/slim/src/router.js
@@ -13,49 +13,70 @@ const homeRoutes = [{
     name: 'home',
     meta: {
         title: 'Home',
-        header: 'Show List'
+        header: 'Show List',
+        topMenu: 'home'
     }
 }, {
     path: '/home/editShow',
-    name: 'editShow'
+    name: 'editShow',
+    meta: {
+        topMenu: 'home'
+    }
 }, {
     path: '/home/displayShow',
-    name: 'show'
+    name: 'show',
+    meta: {
+        topMenu: 'home'
+    }
 }, {
     path: '/home/snatchSelection',
-    name: 'snatchSelection'
+    name: 'snatchSelection',
+    meta: {
+        topMenu: 'home'
+    }
 }, {
     path: '/home/testRename',
     name: 'testRename',
     meta: {
         title: 'Preview Rename',
-        header: 'Preview Rename'
+        header: 'Preview Rename',
+        topMenu: 'home'
     }
 }, {
     path: '/home/postprocess',
     name: 'postprocess',
     meta: {
         title: 'Manual Post-Processing',
-        header: 'Manual Post-Processing'
+        header: 'Manual Post-Processing',
+        topMenu: 'home'
     }
 }, {
     path: '/home/status',
     name: 'status',
     meta: {
-        title: 'Status'
+        title: 'Status',
+        topMenu: 'system'
     }
 }, {
     path: '/home/restart',
     name: 'restart',
     meta: {
         title: 'Restarting...',
-        header: 'Performing Restart'
+        header: 'Performing Restart',
+        topMenu: 'system'
     }
 }, {
     path: '/home/shutdown',
     name: 'shutdown',
     meta: {
-        header: 'Shutting down'
+        header: 'Shutting down',
+        topMenu: 'system'
+    }
+}, {
+    path: '/home/update',
+    name: 'update',
+    meta: {
+        topMenu: 'system'
     }
 }];
 
@@ -65,6 +86,7 @@ const configRoutes = [{
     meta: {
         title: 'Help & Info',
         header: 'Medusa Configuration',
+        topMenu: 'config',
         converted: true
     },
     component: Config
@@ -73,35 +95,40 @@ const configRoutes = [{
     name: 'configAnime',
     meta: {
         title: 'Config - Anime',
-        header: 'Anime'
+        header: 'Anime',
+        topMenu: 'config'
     }
 }, {
     path: '/config/backuprestore',
     name: 'configBackupRestore',
     meta: {
         title: 'Config - Backup/Restore',
-        header: 'Backup/Restore'
+        header: 'Backup/Restore',
+        topMenu: 'config'
     }
 }, {
     path: '/config/general',
     name: 'configGeneral',
     meta: {
         title: 'Config - General',
-        header: 'General Configuration'
+        header: 'General Configuration',
+        topMenu: 'config'
     }
 }, {
     path: '/config/notifications',
     name: 'configNotifications',
     meta: {
         title: 'Config - Notifications',
-        header: 'Notifications'
+        header: 'Notifications',
+        topMenu: 'config'
     }
 }, {
     path: '/config/postProcessing',
     name: 'configPostProcessing',
     meta: {
         title: 'Config - Post Processing',
-        header: 'Post Processing'
+        header: 'Post Processing',
+        topMenu: 'config'
     },
     component: ConfigPostProcessing
 }, {
@@ -109,21 +136,24 @@ const configRoutes = [{
     name: 'configSearchProviders',
     meta: {
         title: 'Config - Providers',
-        header: 'Search Providers'
+        header: 'Search Providers',
+        topMenu: 'config'
     }
 }, {
     path: '/config/search',
     name: 'configSearchSettings',
     meta: {
         title: 'Config - Episode Search',
-        header: 'Search Settings'
+        header: 'Search Settings',
+        topMenu: 'config'
     }
 }, {
     path: '/config/subtitles',
     name: 'configSubtitles',
     meta: {
         title: 'Config - Subtitles',
-        header: 'Subtitles'
+        header: 'Subtitles',
+        topMenu: 'config'
     }
 }];
 
@@ -133,6 +163,7 @@ const addShowRoutes = [{
     meta: {
         title: 'Add Shows',
         header: 'Add Shows',
+        topMenu: 'home',
         converted: true
     },
     component: AddShows
@@ -141,31 +172,38 @@ const addShowRoutes = [{
     name: 'addExistingShows',
     meta: {
         title: 'Add Existing Shows',
-        header: 'Add Existing Shows'
+        header: 'Add Existing Shows',
+        topMenu: 'home'
     }
 }, {
     path: '/addShows/newShow',
     name: 'addNewShow',
     meta: {
         title: 'Add New Show',
-        header: 'Add New Show'
+        header: 'Add New Show',
+        topMenu: 'home'
     }
 }, {
     path: '/addShows/trendingShows',
-    name: 'addTrendingShows'
+    name: 'addTrendingShows',
+    meta: {
+        topMenu: 'home'
+    }
 }, {
     path: '/addShows/popularShows',
     name: 'addPopularShows',
     meta: {
         title: 'Popular Shows',
-        header: 'Popular Shows'
+        header: 'Popular Shows',
+        topMenu: 'home'
     }
 }, {
     path: '/addShows/popularAnime',
     name: 'addPopularAnime',
     meta: {
         title: 'Popular Anime Shows',
-        header: 'Popular Anime Shows'
+        header: 'Popular Anime Shows',
+        topMenu: 'home'
     }
 }];
 
@@ -183,7 +221,8 @@ const addRecommendedRoute = {
     name: 'addRecommended',
     meta: {
         title: 'Add Recommended Shows',
-        header: 'Add Recommended Shows'
+        header: 'Add Recommended Shows',
+        topMenu: 'home'
     },
     component: AddRecommended
 };
@@ -193,7 +232,8 @@ const scheduleRoute = {
     name: 'schedule',
     meta: {
         title: 'Schedule',
-        header: 'Schedule'
+        header: 'Schedule',
+        topMenu: 'schedule'
     }
 };
 
@@ -202,7 +242,8 @@ const historyRoute = {
     name: 'history',
     meta: {
         title: 'History',
-        header: 'History'
+        header: 'History',
+        topMenu: 'history'
     }
 };
 
@@ -211,55 +252,63 @@ const manageRoutes = [{
     name: 'manage',
     meta: {
         title: 'Mass Update',
-        header: 'Mass Update'
+        header: 'Mass Update',
+        topMenu: 'manage'
     }
 }, {
     path: '/manage/backlogOverview',
     name: 'manageBacklogOverview',
     meta: {
         title: 'Backlog Overview',
-        header: 'Backlog Overview'
+        header: 'Backlog Overview',
+        topMenu: 'manage'
     }
 }, {
     path: '/manage/episodeStatuses',
     name: 'manageEpisodeOverview',
     meta: {
         title: 'Episode Overview',
-        header: 'Episode Overview'
+        header: 'Episode Overview',
+        topMenu: 'manage'
     }
 }, {
     path: '/manage/failedDownloads',
     name: 'manageFailedDownloads',
     meta: {
         title: 'Failed Downloads',
-        header: 'Failed Downlaods'
+        header: 'Failed Downlaods',
+        topMenu: 'manage'
     }
 }, {
     path: '/manage/manageSearches',
     name: 'manageManageSearches',
     meta: {
         title: 'Manage Searches',
-        header: 'Manage Searches'
+        header: 'Manage Searches',
+        topMenu: 'manage'
     }
 }, {
     path: '/manage/massEdit',
     name: 'manageMassEdit',
     meta: {
-        title: 'Mass Edit'
+        title: 'Mass Edit',
+        topMenu: 'manage'
     }
 }, {
     path: '/manage/subtitleMissed',
     name: 'manageSubtitleMissed',
     meta: {
         title: 'Missing Subtitles',
-        header: 'Missing Subtitles'
+        header: 'Missing Subtitles',
+        topMenu: 'manage'
     }
 }, {
     path: '/manage/subtitleMissedPP',
     name: 'manageSubtitleMissedPP',
     meta: {
         title: 'Missing Subtitles in Post-Process folder',
-        header: 'Missing Subtitles in Post-Process folder'
+        header: 'Missing Subtitles in Post-Process folder',
+        topMenu: 'manage'
     }
 }];
 
@@ -267,14 +316,16 @@ const errorLogsRoutes = [{
     path: '/errorlogs',
     name: 'errorlogs',
     meta: {
-        title: 'Logs & Errors'
+        title: 'Logs & Errors',
+        topMenu: 'system'
     }
 }, {
     path: '/errorlogs/viewlog',
     name: 'viewlog',
     meta: {
         title: 'Logs',
-        header: 'Log File'
+        header: 'Log File',
+        topMenu: 'system'
     }
 }];
 
@@ -283,7 +334,8 @@ const newsRoute = {
     name: 'news',
     meta: {
         title: 'News',
-        header: 'News'
+        header: 'News',
+        topMenu: 'system'
     }
 };
 
@@ -292,7 +344,8 @@ const changesRoute = {
     name: 'changes',
     meta: {
         title: 'Changelog',
-        header: 'Changelog'
+        header: 'Changelog',
+        topMenu: 'system'
     }
 };
 
@@ -300,7 +353,8 @@ const ircRoute = {
     path: '/IRC',
     name: 'IRC',
     meta: {
-        title: 'IRC'
+        title: 'IRC',
+        topMenu: 'system'
     },
     component: IRC
 };